Move refreshDebouncer to debounce package
